+++ examples/minimal/packages/contracts/types/ethers-contracts/IWorld.ts
@@ -69,6 +69,7 @@ export interface IWorldInterface extends utils.Interface {
     "setRecord(bytes32,bytes32[],bytes,bytes32)": FunctionFragment;
     "staticArrayBytesStruct(tuple[1])": FunctionFragment;
     "staticArrayStringStruct(tuple[1])": FunctionFragment;
+    "transferOwnership(bytes16,address)": FunctionFragment;
     "updateInField(bytes32,bytes32[],uint8,uint256,bytes,bytes32)": FunctionFragment;
     "willRevert()": FunctionFragment;
   };
@@ -106,6 +107,7 @@ export interface IWorldInterface extends utils.Interface {
       | "setRecord"
       | "staticArrayBytesStruct"
       | "staticArrayStringStruct"
+      | "transferOwnership"
       | "updateInField"
       | "willRevert"
   ): FunctionFragment;
@@ -303,6 +305,10 @@ export interface IWorldInterface extends utils.Interface {
     values: [[StringStructStruct]]
   ): string;
   encodeFunctionData(
+    functionFragment: "transferOwnership",
+    values: [PromiseOrValue<BytesLike>, PromiseOrValue<string>]
+  ): string;
+  encodeFunctionData(
     functionFragment: "updateInField",
     values: [
       PromiseOrValue<BytesLike>,
@@ -419,6 +425,10 @@ export interface IWorldInterface extends utils.Interface {
   ): Result;
   decodeFunctionResult(
     functionFragment: "staticArrayStringStruct",
+    data: BytesLike
+  ): Result;
+  decodeFunctionResult(
+    functionFragment: "transferOwnership",
     data: BytesLike
   ): Result;
   decodeFunctionResult(
@@ -731,6 +741,12 @@ export interface IWorld extends BaseContract {
       overrides?: Overrides & { from?: PromiseOrValue<string> }
     ): Promise<ContractTransaction>;
 
+    transferOwnership(
+      namespace: PromiseOrValue<BytesLike>,
+      newOwner: PromiseOrValue<string>,
+      overrides?: Overrides & { from?: PromiseOrValue<string> }
+    ): Promise<ContractTransaction>;
+
     updateInField(
       table: PromiseOrValue<BytesLike>,
       key: PromiseOrValue<BytesLike>[],
@@ -951,6 +967,12 @@ export interface IWorld extends BaseContract {
     overrides?: Overrides & { from?: PromiseOrValue<string> }
   ): Promise<ContractTransaction>;
 
+  transferOwnership(
+    namespace: PromiseOrValue<BytesLike>,
+    newOwner: PromiseOrValue<string>,
+    overrides?: Overrides & { from?: PromiseOrValue<string> }
+  ): Promise<ContractTransaction>;
+
   updateInField(
     table: PromiseOrValue<BytesLike>,
     key: PromiseOrValue<BytesLike>[],
@@ -1166,6 +1188,12 @@ export interface IWorld extends BaseContract {
 
     staticArrayStringStruct(
       arg0: [StringStructStruct],
+      overrides?: CallOverrides
+    ): Promise<void>;
+
+    transferOwnership(
+      namespace: PromiseOrValue<BytesLike>,
+      newOwner: PromiseOrValue<string>,
       overrides?: CallOverrides
     ): Promise<void>;
 
@@ -1436,6 +1464,12 @@ export interface IWorld extends BaseContract {
       overrides?: Overrides & { from?: PromiseOrValue<string> }
     ): Promise<BigNumber>;
 
+    transferOwnership(
+      namespace: PromiseOrValue<BytesLike>,
+      newOwner: PromiseOrValue<string>,
+      overrides?: Overrides & { from?: PromiseOrValue<string> }
+    ): Promise<BigNumber>;
+
     updateInField(
       table: PromiseOrValue<BytesLike>,
       key: PromiseOrValue<BytesLike>[],
@@ -1654,6 +1688,12 @@ export interface IWorld extends BaseContract {
 
     staticArrayStringStruct(
       arg0: [StringStructStruct],
+      overrides?: Overrides & { from?: PromiseOrValue<string> }
+    ): Promise<PopulatedTransaction>;
+
+    transferOwnership(
+      namespace: PromiseOrValue<BytesLike>,
+      newOwner: PromiseOrValue<string>,
       overrides?: Overrides & { from?: PromiseOrValue<string> }
     ): Promise<PopulatedTransaction>;