+++ packages/recs/jest.config.js
@@ -3,7 +3,7 @@
 export default {
   preset: "ts-jest",
   testEnvironment: "node",
-  roots: ["tests"],
+  roots: ["src"],
   moduleNameMapper: {
     // jest can't handle esm imports, so we import the typescript source instead
     "^@latticexyz/common$": "<rootDir>/../common/src/index.ts",