+++ packages/store-sync/src/postgres/getSchema.ts
@@ -1,0 +1,6 @@
+import { PgTable } from "drizzle-orm/pg-core";
+
+// TODO: PR to drizzle to expose `getSchema` like `getTableName`
+export function getSchema(table: PgTable<any>): string | undefined {
+  return (table as any)[Symbol.for("drizzle:Schema")];
+}