+++ packages/store-sync/src/postgres/getTableKey.ts
@@ -1,0 +1,6 @@
+import { getAddress } from "viem";
+import { Table } from "../common";
+
+export function getTableKey(table: Pick<Table, "address" | "namespace" | "name">): string {
+  return `${getAddress(table.address)}:${table.namespace}:${table.name}`;
+}