+++ packages/world/src/World.sol
@@ -11,7 +11,9 @@ import { System } from "./System.sol";
 import { ResourceSelector } from "./ResourceSelector.sol";
 import { ROOT_NAMESPACE, ROOT_NAME } from "./constants.sol";
 import { AccessControl } from "./AccessControl.sol";
-import { Call } from "./Call.sol";
+import { SystemCall } from "./SystemCall.sol";
+import { WorldContextProvider } from "./WorldContext.sol";
+import { revertWithBytes } from "./revertWithBytes.sol";
 
 import { NamespaceOwner } from "./tables/NamespaceOwner.sol";
 import { InstalledModules } from "./tables/InstalledModules.sol";
@@ -47,12 +49,10 @@ contract World is StoreRead, IStoreData, IWorldKernel {
   function installRootModule(IModule module, bytes memory args) public {
     AccessControl.requireOwnerOrSelf(ROOT_NAMESPACE, msg.sender);
 
-    Call.withSender({
+    WorldContextProvider.delegatecallWithContextOrRevert({
       msgSender: msg.sender,
       target: address(module),
-      funcSelectorAndArgs: abi.encodeWithSelector(IModule.install.selector, args),
-      delegate: true, // The module is delegate called so it can edit any table
-      value: 0
+      funcSelectorAndArgs: abi.encodeWithSelector(IModule.install.selector, args)
     });
 
     // Register the module in the InstalledModules table
@@ -176,50 +176,7 @@ contract World is StoreRead, IStoreData, IWorldKernel {
     bytes32 resourceSelector,
     bytes memory funcSelectorAndArgs
   ) external payable virtual returns (bytes memory) {
-    return _call(resourceSelector, funcSelectorAndArgs, msg.value);
-  }
-
-  /**
-   * Call the system at the given namespace and name and pass the given value.
-   * If the system is not public, the caller must have access to the namespace or name.
-   */
-  function _call(
-    bytes32 resourceSelector,
-    bytes memory funcSelectorAndArgs,
-    uint256 value
-  ) internal virtual returns (bytes memory data) {
-    // Load the system data
-    (address systemAddress, bool publicAccess) = Systems.get(resourceSelector);
-
-    // Check if the system exists
-    if (systemAddress == address(0)) revert ResourceNotFound(resourceSelector.toString());
-
-    // Allow access if the system is public or the caller has access to the namespace or name
-    if (!publicAccess) AccessControl.requireAccess(resourceSelector, msg.sender);
-
-    // Get system hooks
-    address[] memory hooks = SystemHooks.get(resourceSelector);
-
-    // Call onBeforeCallSystem hooks (before calling the system)
-    for (uint256 i; i < hooks.length; i++) {
-      ISystemHook hook = ISystemHook(hooks[i]);
-      hook.onBeforeCallSystem(msg.sender, systemAddress, funcSelectorAndArgs);
-    }
-
-    // Call the system and forward any return data
-    data = Call.withSender({
-      msgSender: msg.sender,
-      target: systemAddress,
-      funcSelectorAndArgs: funcSelectorAndArgs,
-      delegate: resourceSelector.getNamespace() == ROOT_NAMESPACE, // Use delegatecall for root systems (= registered in the root namespace)
-      value: value
-    });
-
-    // Call onAfterCallSystem hooks (after calling the system)
-    for (uint256 i; i < hooks.length; i++) {
-      ISystemHook hook = ISystemHook(hooks[i]);
-      hook.onAfterCallSystem(msg.sender, systemAddress, funcSelectorAndArgs);
-    }
+    return SystemCall.callWithHooksOrRevert(msg.sender, resourceSelector, funcSelectorAndArgs, msg.value);
   }
 
   /************************************************************************
@@ -244,8 +201,10 @@ contract World is StoreRead, IStoreData, IWorldKernel {
     // Replace function selector in the calldata with the system function selector
     bytes memory callData = Bytes.setBytes4(msg.data, 0, systemFunctionSelector);
 
-    // Call the function and forward the calldata and returndata
-    bytes memory returnData = _call(resourceSelector, callData, msg.value);
+    // Call the function and forward the call data
+    bytes memory returnData = SystemCall.callWithHooksOrRevert(msg.sender, resourceSelector, callData, msg.value);
+
+    // If the call was successful, return the return data
     assembly {
       return(add(returnData, 0x20), mload(returnData))
     }