+++ packages/world/src/modules/uniqueentity/UniqueEntityModule.sol
@@ -4,7 +4,7 @@ pragma solidity >=0.8.0;
 import { IBaseWorld } from "../../interfaces/IBaseWorld.sol";
 import { IModule } from "../../interfaces/IModule.sol";
 
-import { WorldContext } from "../../WorldContext.sol";
+import { WorldContextConsumer } from "../../WorldContext.sol";
 import { ResourceSelector } from "../../ResourceSelector.sol";
 
 import { UniqueEntity } from "./tables/UniqueEntity.sol";
@@ -16,7 +16,7 @@ import { NAMESPACE, MODULE_NAME, SYSTEM_NAME, TABLE_NAME } from "./constants.sol
  * This module creates a table that stores a nonce, and
  * a public system that returns an incremented nonce each time.
  */
-contract UniqueEntityModule is IModule, WorldContext {
+contract UniqueEntityModule is IModule, WorldContextConsumer {
   // Since the UniqueEntitySystem only exists once per World and writes to
   // known tables, we can deploy it once and register it in multiple Worlds.
   UniqueEntitySystem immutable uniqueEntitySystem = new UniqueEntitySystem();