+++ packages/common/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 2.0.0-next.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @latticexyz/schema-type@2.0.0-next.4
+
 ## 2.0.0-next.3
 
 ### Minor Changes